change flow usecase

+++ domain/src/main/java/com/ptut/domain/model/MovieDomain.kt
@@ -7,7 +7,7 @@ import javax.inject.Inject
 data class MovieDomain(
     val adult: Boolean,
     val backdropPath: String,
-    val genreIds: List<Int>,
+    val genreIds: List<Long>,
     val id: Long,
     val originalLanguage: String,
     val originalTitle: String,
